+++ ComponentFramework/tiny_obj_loader.h
@@ -1662,7 +1662,7 @@ namespace tinyobj {
         std::string token;
 
         bool escaping = false;
-        for (int i = 0; i < s.size(); ++i) {
+        for (size_t i = 0; i < s.size(); ++i) {
             char ch = s[i];
             if (escaping) {
                 escaping = false;
